refactor: 클래스의 책임별로 디렉터리 분리
domain: 현실문제의 의사결정을 담당하는 클래스
sevice: 입력 제공, 결과표시등 의사결정을 돕는 클래스
view: 화면 표시 기능을 담당하는 클래스
