+++ src/view/LottoGameView.js
